// Herbalife.Stock.Web/ClientApp/src/app/shared/models/common.enum.ts
export enum enumAccessOnPage {  
  ManageUser = 'ManageUser',
  Country='Country',
  EntityPermission = 'EntityPermission',
  Location = 'Location', 
  Locator = 'Locator',  
  
  UserGroup = 'UserGroup',
  AssignUser = 'AssignUser',
  
  UserGroupEntityPermission = 'UserGroupEntityPermission',
  
  DataSetup = 'DataSetup',
  
  AssignUsersToUserGroup = 'AssignUsersToUserGroup',
  EntityMaster = 'EntityMaster',
  AssignEntityToUser = 'AssignEntityToUser',
  Home = 'Home',
  EntityReport = 'EntityReport',
  SummaryReport = 'SummaryReport',
  KeyMetric = 'KeyMetric',
  TabTeam = 'TabTeam',
  TopLeaders = 'TopLeaders',
  TopRoyalityPerformer = 'TopRoyalityPerformer',
  QualifyingTabTeam = 'QualifyingTabTeam',
  QualifyingAwt = 'QualifyingAwt',
  VpPopulation = 'VpPopulation',
  ProductPenetration = 'ProductPenetration',
  NutritionClubReport = 'NutritionClubReport',
  KoreaBIHomepage = 'KoreaBIHomepage',
  DeepDiveNote='DeepDiveNote',
  MyReports='MyReports',
  
  
}
export enum enumCheckbox {
  Create = 'Create',
  Read = 'Read',
  Update = 'Update',
  Delete = 'Delete',
  actionCreare = 'C',
  actionRead = 'R',
  actionUpdate = 'U',
  actionDelete = 'D'
}